Move report issue function to issues library
