+++ apps/prairielearn/src/lib/question-submission.ts
@@ -6,7 +6,6 @@ import { HttpStatusError } from '@prairielearn/error';
 import { validateVariantAgainstQuestion } from '../models/variant.js';
 
 import { saveAndGradeSubmission, saveSubmission } from './grading.js';
-import { insertIssue } from './issues.js';
 
 export async function processSubmission(
   req: Request,
@@ -76,41 +75,4 @@ export async function processSubmission(
   } else {
     throw new HttpStatusError(400, `unknown __action: ${req.body.__action}`);
   }
-}
-
-export async function reportIssueFromForm(
-  req: Request,
-  res: Response,
-  studentSubmission = false,
-): Promise<string> {
-  if (studentSubmission && !res.locals.assessment.allow_issue_reporting) {
-    throw new HttpStatusError(403, 'Issue reporting not permitted for this assessment');
-  }
-  const description = req.body.description;
-  if (typeof description !== 'string' || description.length === 0) {
-    throw new HttpStatusError(400, 'A description of the issue must be provided');
-  }
-
-  const variant = await validateVariantAgainstQuestion(
-    req.body.__variant_id,
-    res.locals.question.id,
-    studentSubmission ? res.locals.instance_question?.id : null,
-  );
-  await insertIssue({
-    variantId: variant.id,
-    studentMessage: description,
-    instructorMessage: `${context}-reported issue`,
-    manuallyReported: true,
-    courseCaused: true,
-    courseData: _.pick(res.locals, [
-      'variant',
-      'question',
-      'course_instance',
-      'course',
-      ...(studentSubmission ? ['instance_question', 'assessment_instance', 'assessment'] : []),
-    ]),
-    systemData: {},
-    authnUserId: res.locals.authn_user.user_id,
-  });
-  return variant.id;
 }
